demo-ecs-fastapi-mysql/src/requirements.txt
fastapi==0.101.0
mysql-connector-python==8.1.0
uvicorn==0.23.2